great/requirements.txt
tensorflow-gpu >= 2.2.0
pyaml
wandb
scikit-learn==0.22.1